requirements.txt
# Core dependencies for Limitless Market Maker
colorama>=0.4.6

# Web3 and blockchain dependencies
web3>=6.0.0
eth-account>=0.9.0

# HTTP requests
requests>=2.28.0

# Environment variables
python-dotenv>=1.0.0

# Mathematical operations
numpy>=1.24.0
scipy>=1.10.0

# Data handling
pandas>=2.0.0

# Async operations
aiohttp>=3.8.0

# Logging and monitoring
structlog>=23.0.0
